requirements.txt
google-genai>=1.0.0
groq>=0.9.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
